requirements.txt
asgiref==3.4.1
astroid==2.6.5
colorama==0.4.4
coverage==4.5.1
Django==3.2.5
django-redis==5.0.0
djangorestframework==3.12.4
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==5.3.0
pylint==2.9.5
pylint-runner==0.5.4
pytz==2018.5
redis==3.5.3
selenium==3.14.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.23
wrapt==1.12.1
